requirements.txt
tqdm
opencv-python
numpy
tensorflow-gpu
dlib
PIL
argparse
imutils
scikit-learn
skimage
